Tidy up of Program.cs for bot
